Refactor computer class by moving intelligent computer functionality into its own class
