+++ lib/computer.rb
@@ -1,6 +1,7 @@
 require './lib/board'
 require './lib/cell'
 require './lib/ship'
+require './lib/intelligent_computer.rb'
 
 class Computer
   attr_reader :board, :ships, :user_board
@@ -9,6 +10,7 @@ class Computer
     @board = computer_board
     @user_board = user_board
     @ships = ships
+    @intelligent_computer = IntelligentComputer.new(user_board,computer_board)
   end
 
   def generate_coordinates(length)
@@ -32,62 +34,16 @@ class Computer
     end
   end
 
-  def select_target
-    if @last_hit && !@direction
-      #hit a ship last turn but doesn't know which direction it is facing
-      available_cells = remove_invalid_cells_string(@last_hit.adjacent_cells).values
-
-
-    elsif @last_hit
-      #knows ship and direction
-      if @direction == :x
-        aligned = except(@last_hit.adjacent_cells,[:up,:down])
-      else
-        aligned = except(@last_hit.adjacent_cells,[:left,:right])
-      end
-      available_cells = remove_invalid_cells_string(aligned).values
-
-    else
-      #knows no locations of ships
-      available_cells = remove_invalid_cells_cell(@user_board.cells).keys
-    end
-
-
-    if !available_cells || available_cells.length == 0
-      if @direction == :x
-        index = 0
-      else
-        index = 1
-      end
-
-      all_available = remove_invalid_cells_cell(@user_board.cells).values
-      available_cells = [cells_on_axis(all_available,index).min_by do |cell|
-        #find cell which is the closest to the last shot
-        determine_distance_between_cells(@last_hit,cell,index)
-      end.coordinate]
-      require "Pry"; binding.pry
-    end
-    available_cells.sample
-
-  end
 
   def turn
-    target = select_target
+    target = @intelligent_computer.select_target()
     cell = @user_board.cells[target]
     if target != nil
       cell.fire_upon
       if !cell.empty? #hit
-        if @last_hit && !@direction
-          if @last_hit.coordinate[0] == target[0]
-            @direction = :x
-          else
-            @direction = :y
-          end
-        end
-        @last_hit = cell
+        @intelligent_computer.analyze_hit(target)
         if cell.ship.sunk?
-          @last_hit = nil
-          @direction = nil
+          @intelligent_computer.analyze_sunk
           puts "I sunk your #{cell.ship.name}!"
         end
         puts "My shot on #{target} was a hit."
@@ -96,44 +52,5 @@ class Computer
       end
     end
     #output result of shot
-  end
-
-  def cells_on_axis(available_cells,index)
-    available_cells.select do |cell|
-      @last_hit.coordinate[index] == cell.coordinate[index]
-    end
-  end
-
-  def remove_invalid_cells_string(available_cells)
-    available_cells.select do |direction, coord|
-      @user_board.cells[coord] != nil && !@user_board.cells[coord].fired_upon?
-    end
-  end
-
-  def remove_invalid_cells_cell(available_cells)
-    available_cells.select do |coord, cell|
-      !cell.fired_upon?
-    end
-  end
-
-  def determine_distance_between_cells(cell1,cell2,index)
-    opp_axis_index = index #the opposite axis to the axis index
-    cell_1_pos = cell1.coordinate[opp_axis_index]
-    cell_2_pos = cell2.coordinate[opp_axis_index]
-    if opp_axis_index == 0 #if this is true, it is looking at letters which need to be converted to nums
-      cell_1_pos = cell_1_pos.ord
-      cell_2_pos = cell_2_pos.ord
-    else
-      cell_1_pos = cell_1_pos.to_i
-      cell_2_pos = cell_2_pos.to_i
-    end
-    (cell_1_pos - cell_2_pos).abs()
-  end
-
-  def except(hash,keys) #I do not know why this isn't included with ruby
-    keys.each do |k|
-      hash.delete(k)
-    end
-    hash
   end
 end
